Some updates

+++ libbarobo/barobo.c
@@ -1,11 +1,11 @@
 #include "barobo.h"
 
-int BR_pose(iMobot_t* iMobot, float angle[4], const char motorMask)
+int BR_pose(iMobot_t* iMobot, double angle[4], const char motorMask)
 {
   return 0;
 }
 
-int BR_move(iMobot_t* iMobot, float angle[4], const char motorMask)
+int BR_move(iMobot_t* iMobot, double angle[4], const char motorMask)
 {
   return 0;
 }
@@ -25,7 +25,7 @@ int BR_isBusy(iMobot_t* iMobot)
   return 0;
 }
 
-int BR_getJointAngles(iMobot_t* iMobot, float angle[4])
+int BR_getJointAngles(iMobot_t* iMobot, double angle[4])
 {
   return 0;
 }
